# 4.x: Improve the performance of ToObservable()